+++ scene-graph/src/lib.rs
@@ -1,7 +1,12 @@
 pub mod scene;
 pub mod utils;
 
+pub mod webgpu;
+pub mod webgl;
+
 pub use scene::*;
 pub use utils::*;
+pub use webgpu::*;
+pub use webgl::*;
 
 pub use generational_arena::*;
